InventoryBomb Command Added
Added config true/false for if messages show for all players on server or just the players it affects.
Added InventoryBomb command that drops and fills inventory based on the item specified in the config or minecraft:dead_bush.
Updated PlayerNudger to nudge what the player is riding on or makes the player stop sleeping before nudging.
Updated SwapIdentity to not drop a second copy of a helm if it chooses the same person to swap with and added swap current health